+++ test/testdraw.c
@@ -167,8 +167,8 @@ static void DrawRects(SDL_Renderer *renderer)
 
         rect.w = (float)SDL_rand(viewport.h / 2);
         rect.h = (float)SDL_rand(viewport.h / 2);
-        rect.x = (float)((SDL_rand(viewport.w * 2) - viewport.w) - (rect.w / 2));
-        rect.y = (float)((SDL_rand(viewport.h * 2) - viewport.h) - (rect.h / 2));
+        rect.x = (SDL_rand(viewport.w * 2) - viewport.w) - (rect.w / 2);
+        rect.y = (SDL_rand(viewport.h * 2) - viewport.h) - (rect.h / 2);
         SDL_RenderFillRect(renderer, &rect);
     }
 }
